requirements.txt
async-generator==1.10
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
cssselect==1.1.0
cssutils==2.4.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.2
keyring==23.5.0
lxml==4.8.0
outcome==1.1.0
Pillow==9.0.1
premailer==3.10.0
pycparser==2.21
pyOpenSSL==22.0.0
python-dotenv==0.19.2
pywin32-ctypes==0.2.0
requests==2.27.1
selenium==4.1.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
wsproto==1.0.0
yagmail==0.15.277
zipp==3.7.0
